Move the Resplitter type definition to resplitter package
