+++ datasets/flwr_datasets/federated_dataset.py
@@ -19,8 +19,8 @@ from typing import Dict, Optional, Tuple, Union
 
 import datasets
 from datasets import Dataset, DatasetDict
-from flwr_datasets.common import Resplitter
 from flwr_datasets.partitioner import Partitioner
+from flwr_datasets.resplitter import Resplitter
 from flwr_datasets.utils import (
     _check_if_dataset_tested,
     _instantiate_partitioners,
