+++ datasets/flwr_datasets/resplitter/resplitter.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 # ==============================================================================
-"""Flower Datasets type definitions."""
+"""Resplitter."""
 
 
 from typing import Callable
